# api-rest-cine/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Api rest en GO de un cine",
        "title": "Cine",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4444",
    "basePath": "/",
    "paths": {
...
        "/api/v1/ticket": {
            "post": {
                "description": "Insertar un nuevo ticket en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Insertar un ticket en la base de datos",
                "responses": {
                    "201": {
                        "description": "Creación exitosa",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error en el retorno",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/tickets": {
            "get": {
                "description": "Obtiene todos los tickets ya sea de un usuario registrado o no",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tickets"
                ],
                "summary": "Obtiene todas las tickets",
                "responses": {
                    "200": {
                        "description": "Exitoso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error en la obtención de tickets",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/usuario": {
            "get": {
                "description": "Hace la petición para obtener un solo usuario por su username",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtiene un item por su usuario",
                "parameters": [
                    {
                        "type": "string",
                        "description": "username del usuario a buscar",
                        "name": "user_p",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Insertar un nuevo usuario en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Insertar un usuario en la base de datos",
                "responses": {
                    "201": {
                        "description": "Retorno exitoso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error en la obtencion de los datos",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/usuario/{id}": {
            "get": {
                "description": "Hace la petición para obtener un solo usuario por su ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtiene un usuario por su correo",
                "operationId": "Busca un usuario por su ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la usuario a buscar",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/usuarios": {
            "get": {
                "description": "Obtención de usuarios que ya están registrados en la base de datos",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtención de usuarios",
                "responses": {
                    "201": {
                        "description": "Retorno exitoso",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Error en la obtencion de los datos",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    }
}